dating-app-backend/src/users/users.controller.ts
import { Controller, Get, Param, Request, UseGuards } from '@nestjs/common';
import { JwtAuthGuard } from '../auth/jwt-auth.guard';
import { UsersService } from './users.service';
import { CreateUserDto } from './dto/user.dto';

@Controller('users')
export class UsersController {
  constructor(private readonly usersService: UsersService) {}

  @UseGuards(JwtAuthGuard)
  @Get('profile')
  async getLoggedInUserProfile(@Request() req: any) {
    const userId = req.user.userId; // Extract userId from the decoded JWT
    return this.usersService.findOneById(userId);
  }
}
